adds test coverage for each create-queries permission, turned back on impersonated tests, and deleted duplicate sandboxed tests
